convert buncha stuff to json from xml

+++ DataModels/ItemStuff/FlotsamData.cs
@@ -7,6 +7,6 @@ namespace DataModels.ItemStuff
     public class FlotsamData : IWeightable
     {
         public string Name { get; set; }
-        public int Weight { get; set; }
+        public byte Weight { get; set; }
     }
 }
